Remove FunctionBreakPoint support
We don't need to support FunctionBreakPoint.
generic BreakPoint can support FunctionBreakPoint and there is one problem with current state.
the breakpoints what is setted in function before program launch are not correctly working.
vscode doesn't send any breakpoints with setFunctionBreakPointsRequest when launching
and vscode send every breakpoints with setBreakPointsRequest even if breakpoint is located in function

Signed-off-by: Seonghyun Kim <sh8281.kim@samsung.com>